+++ web/src/components/ZSelect/index.js
@@ -1,0 +1,4 @@
+import ZSelect from './ZSelect';
+
+export { ZSelect };
+export default ZSelect;