+++ web/src/components/index.js
@@ -30,3 +30,5 @@ export { ZLangPicker } from './ZLangPicker';
 export { ZWebsiteNav } from './ZWebsiteNav';
 export { ZLink } from './ZLink';
 export { ZTag } from './ZTag';
+export { ZSelect } from './ZSelect';
+export { ZTextarea } from './ZTextarea';